requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0